requirements/base.txt
django==2.0.3
django-environ==0.4.4
boto3==1.6.16
django-model-utils==3.1.1
kombu==4.1.0
psycopg2cffi==2.7.7
python-dateutil==2.7.2
djangorestframework==3.7.7
gunicorn==19.7.1
django-polymorphic==2.0.2
django-rest-polymorphic==0.1.4
drf-yasg==1.6.0
